requirements.txt
transformers
pytorch
ipykernel
datasets
peft
opacus
evaluate